Moved files, removed TODO, adjusted debounce time
