+++ grunt/tasks/jasmine.js
@@ -8,7 +8,6 @@ var defaultOptions = {
     // Load & install the source-map-support lib (get proper stack traces from inlined source-maps)
     'node_modules/source-map-support/browser-source-map-support.js',
     'test/install-source-map-support.js',
-    // TODO: Use a different (company generated one) API key?
     'http://maps.googleapis.com/maps/api/js?key=AIzaSyA4KzmztukvT7C49NSlzWkz75Xg3J_UyFI'
   ]
 };
