+++ src/geo/geometry-views/gmaps/multi-point-view.js
@@ -1,4 +1,4 @@
-var MultiGeometryViewBase = require('../../geometry-views/base/multi-geometry-view-base');
+var MultiGeometryViewBase = require('../base/multi-geometry-view-base');
 var Pointview = require('./point-view');
 
 var MultiPolygonView = MultiGeometryViewBase.extend({
