+++ src/geo/geometry-views/gmaps/multi-polygon-view.js
@@ -1,4 +1,4 @@
-var MultiGeometryViewBase = require('../../geometry-views/base/multi-geometry-view-base');
+var MultiGeometryViewBase = require('../base/multi-geometry-view-base');
 var PolygonView = require('./polygon-view');
 
 var MultiPolygonView = MultiGeometryViewBase.extend({
