+++ src/geo/geometry-views/leaflet/multi-point-view.js
@@ -1,4 +1,4 @@
-var MultiGeometryViewBase = require('../../geometry-views/base/multi-geometry-view-base');
+var MultiGeometryViewBase = require('../base/multi-geometry-view-base');
 var PointView = require('./point-view');
 
 var MultiPointView = MultiGeometryViewBase.extend({
