+++ test/spec/geo/geometry-views/gmaps/multi-point-view-spec.js
@@ -1,5 +1,5 @@
 var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
-var GMapsMultiPointView = require('../../../../../src/geo/gmaps/geometries/multi-point-view.js');
+var GMapsMultiPointView = require('../../../../../src/geo/geometry-views/gmaps/multi-point-view.js');
 var SharedTestsForMultiPointViews = require('../shared-tests-for-multi-point-views');
 
 describe('src/geo/geometry-views/gmaps/multi-point-view.js', function () {
