+++ test/spec/geo/geometry-views/gmaps/multi-polyline-view-spec.js
@@ -1,5 +1,5 @@
 var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
-var GMapsMultiPolylineView = require('../../../../../src/geo/gmaps/geometries/multi-polyline-view.js');
+var GMapsMultiPolylineView = require('../../../../../src/geo/geometry-views/gmaps/multi-polyline-view.js');
 var SharedTestsForMultiPolylineViews = require('../shared-tests-for-multi-polyline-views');
 
 describe('src/geo/geometry-views/gmaps/multi-polygon-view.js', function () {
