+++ test/spec/geo/geometry-views/gmaps/point-view-spec.js
@@ -1,5 +1,5 @@
 var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
-var GMapsPointView = require('../../../../../src/geo/gmaps/geometries/point-view.js');
+var GMapsPointView = require('../../../../../src/geo/geometry-views/gmaps/point-view.js');
 var SharedTestsForPointViews = require('../shared-tests-for-point-views');
 
 describe('src/geo/geometry-views/gmaps/point-view.js', function () {
