+++ test/spec/geo/geometry-views/gmaps/polygon-view-spec.js
@@ -1,5 +1,5 @@
 var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
-var GMapsPolygonView = require('../../../../../src/geo/gmaps/geometries/polygon-view.js');
+var GMapsPolygonView = require('../../../../../src/geo/geometry-views/gmaps/polygon-view.js');
 var SharedTestsForPolygonViews = require('../shared-tests-for-polygon-views');
 
 describe('src/geo/geometry-views/gmaps/polygon-view.js', function () {
