+++ test/spec/geo/geometry-views/gmaps/polyline-view-spec.js
@@ -1,5 +1,5 @@
 var GMapsMapView = require('../../../../../src/geo/gmaps/gmaps-map-view.js');
-var GMapsPolylineView = require('../../../../../src/geo/gmaps/geometries/polyline-view.js');
+var GMapsPolylineView = require('../../../../../src/geo/geometry-views/gmaps/polyline-view.js');
 var SharedTestsForPolylineViews = require('../shared-tests-for-polyline-views');
 
 describe('src/geo/geometry-views/gmaps/polyline-view.js', function () {
