+++ test/spec/geo/geometry-views/leaflet/multi-point-view-spec.js
@@ -1,5 +1,5 @@
 var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
-var LeafletMultiPointView = require('../../../../../src/geo/leaflet/geometries/multi-point-view.js');
+var LeafletMultiPointView = require('../../../../../src/geo/geometry-views/leaflet/multi-point-view.js');
 var SharedTestsForMultiPointViews = require('../shared-tests-for-multi-point-views');
 
 describe('src/geo/geometry-views/leaflet/multi-point-view.js', function () {
