+++ test/spec/geo/geometry-views/leaflet/multi-polygon-view-spec.js
@@ -1,5 +1,5 @@
 var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
-var LeafletMultiPolygonView = require('../../../../../src/geo/leaflet/geometries/multi-polygon-view.js');
+var LeafletMultiPolygonView = require('../../../../../src/geo/geometry-views/leaflet/multi-polygon-view.js');
 var SharedTestsForMultiPolygonViews = require('../shared-tests-for-multi-polygon-views');
 
 describe('src/geo/geometry-views/leaflet/multi-polygon-view.js', function () {
