+++ test/spec/geo/geometry-views/leaflet/multi-polyline-view-spec.js
@@ -1,5 +1,5 @@
 var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
-var LeafletMultiPolylineView = require('../../../../../src/geo/leaflet/geometries/multi-polyline-view.js');
+var LeafletMultiPolylineView = require('../../../../../src/geo/geometry-views/leaflet/multi-polyline-view.js');
 var SharedTestsForMultiPolylineViews = require('../shared-tests-for-multi-polyline-views');
 
 describe('src/geo/geometry-views/leaflet/multi-polyline-view.js', function () {
