+++ test/spec/geo/geometry-views/leaflet/point-view-spec.js
@@ -1,5 +1,5 @@
 var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
-var LeafletPointView = require('../../../../../src/geo/leaflet/geometries/point-view.js');
+var LeafletPointView = require('../../../../../src/geo/geometry-views/leaflet/point-view.js');
 var SharedTestsForPointViews = require('../shared-tests-for-point-views');
 
 describe('src/geo/geometry-views/leaflet/point-view.js', function () {
