+++ test/spec/geo/geometry-views/leaflet/polyline-view-spec.js
@@ -1,5 +1,5 @@
 var LeafletMapView = require('../../../../../src/geo/leaflet/leaflet-map-view.js');
-var LeafletPolylineView = require('../../../../../src/geo/leaflet/geometries/polyline-view.js');
+var LeafletPolylineView = require('../../../../../src/geo/geometry-views/leaflet/polyline-view.js');
 var SharedTestsForPolylineViews = require('../shared-tests-for-polyline-views');
 
 describe('src/geo/geometry-views/leaflet/polyline-view.js', function () {
